+++ integration/cmd_app_test.go
@@ -1,11 +1,9 @@
 package integration_test
 
 import (
-	"context"
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/starport/starport/pkg/cmdrunner/step"
@@ -51,35 +49,6 @@ func TestGenerateAnAppWithCosmWasmAndVerify(t *testing.T) {
 	env.EnsureAppIsSteady(path)
 }
 
-func TestServeAppWithCosmWasm(t *testing.T) {
-	t.Parallel()
-
-	var (
-		env     = newEnv(t)
-		apath   = env.Scaffold("blog", Launchpad)
-		servers = env.RandomizeServerPorts(apath)
-	)
-
-	env.Exec("add CosmWasm module",
-		step.New(
-			step.Exec("starport", "module", "import", "wasm"),
-			step.Workdir(apath),
-		),
-	)
-
-	var (
-		ctx, cancel       = context.WithTimeout(env.Ctx(), time.Minute*5)
-		isBackendAliveErr error
-	)
-	go func() {
-		defer cancel()
-		isBackendAliveErr = env.IsAppServed(ctx, servers)
-	}()
-	env.Serve("should serve with CosmWasm", apath, ExecCtx(ctx))
-
-	require.NoError(t, isBackendAliveErr, "app cannot get online in time")
-}
-
 func TestGenerateAnAppWithEmptyModuleAndVerify(t *testing.T) {
 	t.Parallel()
 
@@ -109,24 +78,6 @@ func TestGenerateAStargateAppWithEmptyModuleAndVerify(t *testing.T) {
 	env.Exec("add CosmWasm module",
 		step.New(
 			step.Exec("starport", "module", "create", "example"),
-			step.Workdir(path),
-		),
-	)
-
-	env.EnsureAppIsSteady(path)
-}
-
-func TestGenerateAnAppWithTypeAndVerify(t *testing.T) {
-	t.Parallel()
-
-	var (
-		env  = newEnv(t)
-		path = env.Scaffold("blog", Stargate)
-	)
-
-	env.Exec("add CosmWasm module",
-		step.New(
-			step.Exec("starport", "type", "user", "email"),
 			step.Workdir(path),
 		),
 	)
