+++ integration/cmd_serve_test.go
@@ -1,0 +1,39 @@
+package integration_test
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+	"github.com/tendermint/starport/starport/pkg/cmdrunner/step"
+)
+
+func TestServeAppWithCosmWasm(t *testing.T) {
+	t.Parallel()
+
+	var (
+		env     = newEnv(t)
+		apath   = env.Scaffold("blog", Launchpad)
+		servers = env.RandomizeServerPorts(apath)
+	)
+
+	env.Exec("add CosmWasm module",
+		step.New(
+			step.Exec("starport", "module", "import", "wasm"),
+			step.Workdir(apath),
+		),
+	)
+
+	var (
+		ctx, cancel       = context.WithTimeout(env.Ctx(), time.Minute*5)
+		isBackendAliveErr error
+	)
+	go func() {
+		defer cancel()
+		isBackendAliveErr = env.IsAppServed(ctx, servers)
+	}()
+	env.Serve("should serve with CosmWasm", apath, ExecCtx(ctx))
+
+	require.NoError(t, isBackendAliveErr, "app cannot get online in time")
+}
